# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Offera Proposal Builder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ProposalStatus =
  (typeof ProposalStatus)[keyof typeof ProposalStatus];

export const ProposalStatus = {
  draft: "draft",
  sent: "sent",
  viewed: "viewed",
  accepted: "accepted",
  declined: "declined",
} as const;

export interface PricingRow {
  id: string;
  description: string;
  quantity: number;
  unitPrice: number;
  total: number;
}

export type ContentBlockType =
  (typeof ContentBlockType)[keyof typeof ContentBlockType];

export const ContentBlockType = {
  heading: "heading",
  text: "text",
  image: "image",
  pricing: "pricing",
  divider: "divider",
} as const;

export interface ContentBlock {
  id: string;
  type: ContentBlockType;
  content?: string;
  rows?: PricingRow[];
  discount?: number;
  vatEnabled?: boolean;
  imageUrl?: string;
  level?: number;
}

export interface ProposalSection {
  id: string;
  title: string;
  blocks: ContentBlock[];
}

export type ProposalBrandingFont =
  (typeof ProposalBrandingFont)[keyof typeof ProposalBrandingFont];

export const ProposalBrandingFont = {
  inter: "inter",
  playfair: "playfair",
  "dm-sans": "dm-sans",
} as const;

export interface ProposalBranding {
  logoUrl?: string;
  accentColor: string;
  font: ProposalBrandingFont;
}

export interface Proposal {
  id: number;
  title: string;
  clientName: string;
  clientEmail?: string;
  status: ProposalStatus;
  totalValue: number;
  publicSlug: string;
  sections: ProposalSection[];
  branding: ProposalBranding;
  personalMessage?: string;
  createdAt: string;
  updatedAt: string;
  lastActivityAt?: string;
}

export interface CreateProposalRequest {
  title: string;
  clientName: string;
  clientEmail?: string;
}

export interface UpdateProposalRequest {
  title?: string;
  clientName?: string;
  clientEmail?: string;
  sections?: ProposalSection[];
  branding?: ProposalBranding;
  totalValue?: number;
}

export interface SendProposalRequest {
  clientEmail: string;
  personalMessage?: string;
}

export type RespondToProposalRequestAction =
  (typeof RespondToProposalRequestAction)[keyof typeof RespondToProposalRequestAction];

export const RespondToProposalRequestAction = {
  accept: "accept",
  decline: "decline",
} as const;

export interface RespondToProposalRequest {
  action: RespondToProposalRequestAction;
}