# lib/api-zod/src/generated/types/contentBlock.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Offera Proposal Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { ContentBlockType } from "./contentBlockType";
import type { PricingRow } from "./pricingRow";

export interface ContentBlock {
  id: string;
  type: ContentBlockType;
  content?: string;
  rows?: PricingRow[];
  discount?: number;
  vatEnabled?: boolean;
  imageUrl?: string;
  level?: number;
}