# lib/api-zod/src/generated/types/proposal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Offera Proposal Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { ProposalBranding } from "./proposalBranding";
import type { ProposalSection } from "./proposalSection";
import type { ProposalStatus } from "./proposalStatus";

export interface Proposal {
  id: number;
  title: string;
  clientName: string;
  clientEmail?: string;
  status: ProposalStatus;
  totalValue: number;
  publicSlug: string;
  sections: ProposalSection[];
  branding: ProposalBranding;
  personalMessage?: string;
  createdAt: Date;
  updatedAt: Date;
  lastActivityAt?: Date;
}