# lib/api-zod/src/generated/types/proposalBranding.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Offera Proposal Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { ProposalBrandingFont } from "./proposalBrandingFont";

export interface ProposalBranding {
  logoUrl?: string;
  accentColor: string;
  font: ProposalBrandingFont;
}