# lib/api-zod/src/generated/types/proposalStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Offera Proposal Builder API
 * OpenAPI spec version: 0.1.0
 */

export type ProposalStatus =
  (typeof ProposalStatus)[keyof typeof ProposalStatus];

export const ProposalStatus = {
  draft: "draft",
  sent: "sent",
  viewed: "viewed",
  accepted: "accepted",
  declined: "declined",
} as const;