# lib/api-zod/src/generated/types/respondToProposalRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Offera Proposal Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { RespondToProposalRequestAction } from "./respondToProposalRequestAction";

export interface RespondToProposalRequest {
  action: RespondToProposalRequestAction;
}