# lib/api-zod/src/generated/types/sendProposalRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Offera Proposal Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface SendProposalRequest {
  clientEmail: string;
  personalMessage?: string;
}