# lib/api-zod/src/generated/types/updateProposalRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Offera Proposal Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { ProposalBranding } from "./proposalBranding";
import type { ProposalSection } from "./proposalSection";

export interface UpdateProposalRequest {
  title?: string;
  clientName?: string;
  clientEmail?: string;
  sections?: ProposalSection[];
  branding?: ProposalBranding;
  totalValue?: number;
}